Add files via upload
Just worked on it really quick. Database is added + change font/ background/forcolor to the table. Error handling for everything. hmmmm. All it needs now is to display the data when clicking a row. I think we also need to change the headers of the table, but this will do for now since I have other homework, lol.